% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_RandSearchAlgorithm.R
\name{adjustRangesFromCentralPoint}
\alias{adjustRangesFromCentralPoint}
\title{Adjust parameter ranges}
\usage{
adjustRangesFromCentralPoint(paramList, centralPoint, neighSizeProp = 0.025)
}
\arguments{
\item{paramList}{An input list containing the parameter ranges}

\item{centralPoint}{A vector containing coordinates (parameter values) for the best point. 
It must have the same size as \code{length(paramList)}.}

\item{neighSizeProp}{Size of the neighbourhood for a given parameter, i.e., a real value contained 
in ]0, 1] used to multiply the range size as: \eqn{neighSize = (max_{range} - min_{range}) \times neighSizeProp}.}
}
\value{
A list object containing the updated ranges.
}
\description{
An auxiliary function used in \link{randomSearchOptim} to adjust ranges around a 
point. This adjustement is required to draw a random set of points in the vicinity of the best 
central point in a given iteration.
}
\examples{

params <- list("thresh"  = c(0.1,0.5), "minsize" = c(10,20))

adjustRangesFromCentralPoint(paramList = params, 
centralPoint = c(0.15, 12), neighSizeProp = 0.025)

}
