% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_CalcSegmentStats.R
\name{aggregateMultiStats}
\alias{aggregateMultiStats}
\title{Aggregate data by segment ID}
\usage{
aggregateMultiStats(x, funs = c("mean", "sd"), na.rm = TRUE)
}
\arguments{
\item{x}{A data.frame or matrix object.}

\item{funs}{Function names (default: \code{c("mean","sd")}).}

\item{na.rm}{Remove NA's? This will be passed to aggregation functions defined 
(default: TRUE).}
}
\description{
An ancillary function used to aggregate data by segment ID (field: SID) for multiple 
functions
}
