% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_PerformanceEval.R
\name{aucSingleClass}
\alias{aucSingleClass}
\title{Calculate the Area Under the ROC Curve (AUC)}
\usage{
aucSingleClass(obs, pred)
}
\arguments{
\item{obs}{Integer vector with observed values (class labels, tipically 0 or 1).}

\item{pred}{Numeric vector with predicted probabilities.}
}
\value{
Numeric. The AUC value calculated.
}
\description{
A  function used to calculate the Area Under the (Receiver Operating Characteristic) 
Curve (AUC) when the predicted output is given in probability. This function applies 
only to single-class problems (e.g., presence/absence of a vegetation type, species).
}
\details{
Uses the \pkg{ROCR} package for calculations.
}
\examples{
obs<-sample(c(0,1),100,replace = TRUE)
pred<-runif(100,0,1)
aucSingleClass(obs,pred)

}
\seealso{
\code{\link[ROCR]{prediction}}, \code{\link[ROCR]{performance}}
}
