% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_VectorOps_DiscrepancyMeasures.R
\name{calcDiscrepancyMeasure_ED2}
\alias{calcDiscrepancyMeasure_ED2}
\title{Calculate ED2 segmentation discrepancy measure}
\usage{
calcDiscrepancyMeasure_ED2(
  refData,
  segmData,
  thresh = 50,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{refData}{Reference polygons data. Either a string containing a path to a vector 
data set or a simple features geometry object (\code{sf}). See details below.}

\item{segmData}{Segmentation polygons data. Either a string containing a path to a vector 
data set or a simple features geometry object (\code{sf}).}

\item{thresh}{A threshold percentage used to select segments for each reference polygon. 
Calculated as: \eqn{\%SegmIntersected = (AreaIntersected / AreaSegment) \times 100}. If 
\%SegmIntersected > thresh then that segment will be selected (default: 50).}

\item{verbose}{Print progress messages? (default: FALSE)}

\item{...}{Further arguments passed to \code{\link[sf]{read_sf}} function.}
}
\value{
A numeric vector with several outputs:        
\itemize{      
    \item \strong{ED2} - The Euclidean Distance measure 2 (Liu et al., 2012);
    \item \strong{PSE} - Potential Segmentation Error;
    \item \strong{NSR} - Number-of-Segments Ratio;
    \item \strong{NrRefPolygons} - Total number of reference polygons; 
    \item \strong{NrOfRefPolsWoCandSegms} - Number of reference polygons without 
    candidate segments;
    \item \strong{NrOfCandSegments} - Number of candidate segments (for which 
    the % of intersected area is above thresh);
    \item \strong{SumUndersegAreaAbsDiff} - Sum of absolute undersegmented 
    area differences;
    \item \strong{TotalAreaRefPols} - Total area for reference polygons (with 
    one or more candidate segments).
}
}
\description{
This function implements the ED2 discrepancy measure used to assess a given 
image segmentation solution. It is based on Liu et al., 2012 with some corrections 
introduced later by Novelli et al., 2017.
}
\details{
A \strong{"reference polygon"} can be considered a bounded geographic region (Castilla, 2003; 
Lang et al., 2010). A basic region is an indivisible unit with respect to certain perspectives 
of human cognition (Castilla and Hay, 2008) and can often be identified easily from its surroundings. 
In this case a reference polygon is considered a geographic object of interest. These reference polygons 
are typically sampled from EO imagery and delineated through manual interpretation or from field surveys. 
The size of a reference polygon is typically greater than that of a pixel in an image and smaller than 
the whole extent of that image.
}
\note{
Compared to AssesSeg tool (from Novelli et al., 2017), this function has some 
slight numerical differences in cases where the number of reference polygons without 
candidate segments is higher than 0. Probably these derive from area calculations or 
specific aspects of each implementation.
}
\references{
Castilla, G., 2003. Object-oriented analysis of remote sensing images for land cover mapping: 
conceptual foundations and a segmentation method to derive a baseline partition for 
classification. PhD Thesis, Polytechnic University of Madrid.              

Castilla, G., Hay, G.J., 2008. Image objects and geographic objects. In: Blaschke, T., Lang, S. (Eds.), 
Object-based Image Analysis: Spatial Concepts for Knowledgedriven Remote Sensing Applications. 
Springer-Verlag, Berling, Heidelberg, pp. 91-110.  

Lang, S., Albrecht, F., Kienberger, S., Tiede, D., 2010. Object validity for operational tasks in 
a policy context. Journal of Spatial Science 55 (1), 9-22. 

Liu, Y., L. Bian, Y. Meng, H. Wang, S. Zhang, Y. Yang, X. Shao, and B. Wang. 2012. 
Discrepancy measures for selecting optimal combination of parameter values in 
object-based image analysis. ISPRS Journal of Photogrammetry and Remote Sensing 
68:144-156.

Novelli, A., M. Aguilar, F. Aguilar, A. Nemmaoui, and E. Tarantino. 2017. AssesSeg - 
A Command Line Tool to Quantify Image Segmentation Quality: A Test Carried Out in 
Southern Spain from Satellite Imagery. Remote Sensing 9:40.
}
