% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_GetTrainData.R
\name{calcStatsFinish}
\alias{calcStatsFinish}
\title{Ancillary function used in getTrainData_ for final processing (by tile)}
\usage{
calcStatsFinish(x, thresh = 0.5)
}
\arguments{
\item{x}{A data.frame from \code{\link{calcStatsStart}}.}

\item{thresh}{A threshold value defining the minimum proportion of the segment ]0, 1] that 
must be covered by a certain class to be considered as a training case. This threshold will 
only apply if \code{x} is a \code{RasterLayer} which means you are using train areas/pixels. 
If you are running a \emph{"single-class"} problem then this threshold only applies to the 
class of interest (coded as 1's). Considering this, if a given segment has a proportion cover 
of that class higher than \code{thresh} then it is considered a train case. In contrast, for 
the background class (coded as 0's), only segments/objects totaly covered by that class are 
considered as train cases.      
If you are running a \emph{"multi-class"} problem then \code{thresh} is applied differently. 
First, the train class is determined by a majority rule then if that class covers more than 
the value specified in \code{thresh} this case is kept in train data otherwise it will be 
filtered out. See also \code{useThresh}.}
}
\value{
A data.frame with final training data.
}
\description{
Performs data aggregation ops necessary to calculate the percentage cover per 
class and segment ID based on initial processing by \code{\link{calcStatsStart}}.
}
