% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_GetTrainData.R
\name{calcStatsStart}
\alias{calcStatsStart}
\title{Ancillary function used in getTrainData_ for initial processing (by tile)}
\usage{
calcStatsStart(x)
}
\arguments{
\item{x}{A data.frame or matrix with two columns named: "SID" (segment ID's) and 
"train" (with train class or NA)}
}
\value{
A data.frame with preliminary training data.
}
\description{
Performs data aggregation ops necessary to calculate the number of pixels per 
class and segment ID.
}
