% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_ExtractFeaturesIndices.R
\name{calculateNDIcombinations}
\alias{calculateNDIcombinations}
\title{Calculate Normalized Difference Index (NDI) combinations}
\usage{
calculateNDIcombinations(
  rst,
  bandNames = NULL,
  subsetBands = NULL,
  getRstStack = TRUE,
  scale10k = FALSE,
  filename = NULL,
  writeSingleBandRaster = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{rst}{An input \code{RasterStack} object with n > 1 layers.}

\item{bandNames}{A character vector defining band names (e.g., c("Bl","Gr","Rd","NIR")).}

\item{subsetBands}{An integer vector defining which bands to be used.}

\item{getRstStack}{Should the function return a final \code{RasterStack} object 
with all NDI combinations? (default: TRUE) If many layers exist in rst then it is advisable 
that this option is set to FALSE otherwise lack of memory problems may happen.}

\item{scale10k}{Should data be scaled/multiplied by a factor of 10000 (default: FALSE). If 
TRUE then the raster data type is set to "INT4S" (otherwise "FLT4S" is used). Used to prevent 
out-of-memory problems.}

\item{filename}{A base filename used to export NDI combinations as a multiband (if 
writeSingleBandRaster=FALSE) or single-band (if otherwise).}

\item{writeSingleBandRaster}{If an output filename is defined should each 
NDI raster be written separetely? (default: FALSE) This is used to save up memory 
by not having to stack up each NDI raster. If writeSingleBandRaster is set to TRUE 
then a filename must be defined. The final filename will have the band indices or names 
as suffix.}

\item{verbose}{Should the function print out progress messages? (default: TRUE)}

\item{...}{Additional arguments passed to}
}
\value{
A \code{RasterStack} object is returned if \code{getRstStack=TRUE}.
}
\description{
Calculates (all or some) NDI combinations based on an input raster stack with n>1 layers. Combinations are 
obtained using the \code{\link[utils]{combn}} function (for combinations of n layers, taken 2 
at a time).
}
