% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_CalcSegmentStats.R
\name{calculateSegmentStats}
\alias{calculateSegmentStats}
\title{Calculate segment statistics for multiple raster features and aggregation functions}
\usage{
calculateSegmentStats(
  rstFeatures,
  rstSegm,
  funs = c("mean", "sd"),
  na.rm = TRUE,
  bylayer = FALSE,
  tiles = NULL,
  subset = NULL,
  progressBar = FALSE
)
}
\arguments{
\item{rstFeatures}{A string defining the path to the raster features or a 
\code{rasterStack} object.}

\item{rstSegm}{A string defining the path to the raster with segment IDs or a 
\code{rasterLayer} object.}

\item{funs}{A character vector with the name(s) of the functions used to aggregate 
data (default: \code{c("mean", "sd")}).}

\item{na.rm}{Boolean defining if NA's should be removed (default: TRUE).}

\item{bylayer}{Calculate statistics layer by layer instead of all at once? (slightly 
increases computation time but spares memory load; default: FALSE).}

\item{tiles}{Number of times to slice the RasterLayer across row 
and column direction. The total number of tiles will be given by: 
\eqn{N_{tiles} = nd^{2}}.}

\item{subset}{A vector defining which segments should be selected.}

\item{progressBar}{Boolean. Show progress bar? (default: FALSE).}
}
\value{
A data frame containing segment ids (first column, SID) and aggregated features. 
The name of the aggregation function will be appended to the name of each raster feature.
}
\description{
This function allows calculating aggregation statistics by segment ID to be used in the 
classification stage. This function employs \code{dtplyr} (based on \code{data.table}) 
and \code{dplyr} verbs for processing.
}
\examples{

library(raster)
library(dplyr)
library(dtplyr)

rstSegm <- raster(nrows=100, ncols=100, xmn=0, xmx=100, 
                  ymn=0, ymx=100)

km <- kmeans(coordinates(rstSegm),100,iter.max = 100)
values(rstSegm) <- km$cluster
   
rstFeatures_1 <- raster(nrows=100, ncols=100, xmn=0, xmx=100, 
ymn=0, ymx=100, res=1, vals=runif(10000))

rstFeatures_2 <- raster(nrows=100, ncols=100, xmn=0, xmx=100, 
ymn=0, ymx=100, res=1, vals=runif(10000))
   
rstStackFeatures <- stack(rstFeatures_1, rstFeatures_2)

calculateSegmentStats(rstStackFeatures, rstSegm, 
                      funs = c("mean", "sd"), na.rm = TRUE)


}
