% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_CheckOptimFuncParameters.R
\name{classificationAlgorithms}
\alias{classificationAlgorithms}
\title{Get classification algorithm codes}
\usage{
classificationAlgorithms()
}
\value{
A vector with codes/acronyms of classification algorithms.
}
\description{
Outputs the codes/acronyms of classification algorithms.
}
