% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_CheckOptimFuncParameters.R
\name{classifierParameters}
\alias{classifierParameters}
\title{Get the evaluation methods names}
\usage{
classifierParameters()
}
\value{
A vector with codes/acronyms of evaluation methods.
}
\description{
Outputs the codes/acronyms of evaluation methods.
}
