% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_UnsupervisedClassification.R
\name{clusteringRaster}
\alias{clusteringRaster}
\title{Perform clustering on a RasterStack object and calculate internal clustering criteria}
\usage{
clusteringRaster(
  inRst,
  k,
  writeRasterData = TRUE,
  outRst,
  getRstStack = TRUE,
  method = "kmeans",
  calcIntCriteria = FALSE,
  crit = c("Silhouette"),
  intCritSampSize = 20000,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{inRst}{An input \code{RasterStack} object. By default all layers/bands will be used to perform the k-means 
clustering.}

\item{k}{An integer vector defining the number of clusters. If more than one value is given then 
\code{kmeansRaster} will run one time for each \code{k} value.}

\item{writeRasterData}{Write clustered raster data to file (i.e., to outRst)? (default:TRUE)}

\item{outRst}{Name of the output raster. For each value in \code{k} the file name will get a suffix 
in this form: \code{"filename_method_nc_k.ext"}.}

\item{getRstStack}{Get an output \code{RasterStack} object with clustering solutions? (default: \code{FALSE}). 
If \code{FALSE} then some memory saving is expected.}

\item{method}{A string defining the method used for clustering. Available options are: 
\itemize{
   \item \strong{\code{"kmeans"}}: for k-means partitioning clustering from stats package (default); 
   \item \strong{\code{"hardcl"}}: for hard-competitive learning; 
   \item \strong{\code{"neuralgas"}}: for neural-gas algorithm both from cclust package, and; 
   \item \strong{\code{"clara"}}: from cluster package.
}}

\item{calcIntCriteria}{Calculate internal clustering criteria? (default: FALSE).}

\item{crit}{A character vector containing the names of the indices to compute. Check available indices 
at \code{\link[clusterCrit]{intCriteria}} or running \code{getCriteriaNames(TRUE)}.}

\item{intCritSampSize}{Number of sample pixels to use for internal criteria calculation (default: 20000). 
Increasing this number will make estimations better however at the expense of more memory.}

\item{verbose}{Print progress messages? (default: TRUE).}

\item{...}{Further parameters passed to each of clustering functions \code{\link[stats]{kmeans}}, 
\code{\link[cclust]{cclust}} or, \code{\link[cluster]{clara}} (for clara, except the following 
paramaters which are pre-defined to: \code{keep.data = FALSE}, and \code{medoids.x = FALSE} to 
save memory and speed-up calculations).}
}
\value{
If \code{getRstStack=TRUE} then a \code{RasterStack} object with all clustering solutions will be returned. 
If \code{getRstStack=TRUE} and \code{calcIntCriteria=TRUE} then an additional matrix named \code{intClustCrit} with 
one row by k value will be returned with the requested clustering indices. If \code{getRstStack=FALSE} then only the matrix 
with clustering indices will be returned.
}
\description{
Runs a clustering/unsupervised learning/classification algorithm on a multi-layer \code{RasterStack} object.      
It also allows calculating internal clustering performance criteria based on function 
\code{\link[clusterCrit]{intCriteria}} using a stratified random sample of output pixels.
}
\examples{

library(raster)

r1 <- raster(nrows=10, ncols=10, vals=rnorm(100))
r2 <- raster(nrows=10, ncols=10, vals=rnorm(100))
r3 <- raster(nrows=10, ncols=10, vals=rnorm(100))

rstStack <- stack(r1, r2, r3)

clusteringRaster(rstStack, k = 2:10, writeRasterData = FALSE, getRstStack = TRUE, 
                 method="kmeans", calcIntCriteria = FALSE, crit = c("Silhouette"), 
                 intCritSampSize = 20000, verbose = TRUE)


}
\seealso{
See \code{\link[clusterCrit]{intCriteria}}, \code{\link[clusterCrit]{getCriteriaNames}} for more details on 
clustering criteria.      
Check out clustering functions at: \code{\link[stats]{kmeans}}, \code{\link[cclust]{cclust}}, and, 
\code{\link[cluster]{clara}}.
}
