% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_ImportConvertData.R
\name{convertToSAGA}
\alias{convertToSAGA}
\title{Convert files to SAGA raster format}
\usage{
convertToSAGA(rst, outRasterPath, verbose = TRUE, ...)
}
\arguments{
\item{rst}{An input \code{RasterLayer}, \code{RasterStack} object or alternatively a character indicating a raster file 
path used for conversion.}

\item{outRasterPath}{The output SAGA raster .sdat file.}

\item{verbose}{Print messages? (default: TRUE)}

\item{...}{Additional parameters passed to \code{\link[raster]{writeRaster}}.}
}
\description{
An auxiliar function used to convert a \code{RasterLayer}, \code{RasterStack} or any raster file (given by a file path) 
into SAGA .sdat/.sgrd raster format.
}
\details{
For \code{RasterStack} objects each individual layer (or band) will be convert to a single .sdat file (this format has 
no support for multiband raster files). Each output file will be given a suffix _b1, _b2, ..., _bn according to the band 
position.
}
