% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_ClassificationFunctions.R
\name{createDataSplits}
\alias{createDataSplits}
\title{Create a data split for classification evaluation}
\usage{
createDataSplits(y, evalMethod = "10FCV", nRounds = 20, trainPerc = 0.8)
}
\arguments{
\item{y}{A vector containing train inputs.}

\item{evalMethod}{A character string defining the evaluation method. The available methods are \code{"10FCV"} (10-fold 
cross-validation; the default), \code{"5FCV"}, (5-fold cross-validation), \code{"HOCV"} (holdout cross-validation with 
the training percentage defined by \code{trainPerc} and the number of rounds defined in \code{nRounds}), and, \code{"OOB"} 
(out-of-bag evaluation; only applicable to random forests algorithm).}

\item{nRounds}{Number of evaluation rounds. This is only used for HOCV method.}

\item{trainPerc}{Proportion of data used for training the algorithm (by deafult eaul to 0.8). This is 
only used for the \code{"HOCV"} method.}
}
\value{
A list like object of class 'dataSplit' to be used in calibrateClassifier() function. Each element on the list 
contains the indices used for training.
}
\description{
An auxiliary/internal function used to create a data split intended to evaluate classification performance. 
Used internally in the \code{\link{calibrateClassifier}} function.
}
\details{
Uses function \link[caret]{createDataPartition} from \pkg{caret} package as the 'workhorse' to generate 
data splits.
}
\examples{
createDataSplits(y=rnorm(100), evalMethod = "10FCV", nRounds = 20, trainPerc = 0.8)

}
