% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_RasterTiling.R
\name{createRasterTiles}
\alias{createRasterTiles}
\title{Create a raster tiles object}
\usage{
createRasterTiles(rst, nd = 2)
}
\arguments{
\item{rst}{A \code{SpatRaster} object.}

\item{nd}{Number of times to slice the SpatRaster across row 
and column direction. The total number of tiles will be given by: 
\eqn{N_{tiles} = nd^{2}}.}
}
\value{
A \code{SOptim.Tiles} object with nd^2 tiles each with the starting 
row and column numbers as well as the number of rows and columns to read.
}
\description{
A function used to cut an input \code{SpatRaster} object into several tiles 
usable by \code{\link[terra]{values}} to read data by chunks.
}
\examples{

library(terra)
r <- rast(ncol = 100, nrow = 100)
rtiles <- createRasterTiles(r, nd = 3)

}
