% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_DataBalancing.R
\name{dataBalancing}
\alias{dataBalancing}
\title{Perform data balancing}
\usage{
dataBalancing(x, method)
}
\arguments{
\item{x}{A data frame or matrix for balancing (by default the two first columns are 
assumed to be the segment ID and the train labels).The response variable of the 
unbalanced dataset (i.e., the column named as "train") must be a binary factor 
where the majority class is coded as 0's and the minority (the class of interest) as 1's.}

\item{method}{A string defining the method to apply. Either "ubOver" for over-sampling 
the minority class or "ubUnder" for under-sampling the majority class.}
}
\value{
A data frame with balanced classes.
}
\description{
An internal wrapper function for performing data balancing for single-class problems. 
Two methods are available via the \code{unbalanced} package: i) over-sampling 
\code{\link[unbalanced]{ubOver}} or, ii) under-sampling \code{\link[unbalanced]{ubUnder}}
}
\examples{

DF <- data.frame(SID=1:100,train=c(rep(0,90),rep(1,10)),matrix(rnorm(1000),10,100))

DF.over<-dataBalancing(x=DF, method="ubOver")
DF.under<-dataBalancing(x=DF,method="ubUnder")

}
