% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_CheckOptimFuncParameters.R
\name{dataBalancingMethods}
\alias{dataBalancingMethods}
\title{Classification algorithm codes}
\usage{
dataBalancingMethods()
}
\value{
A vector with codes/acronyms of classification algorithms.
}
\description{
Outputs the codes/acronyms of classification algorithms.
}
