% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_AuxFunctions.R
\name{doCleanUpActions}
\alias{doCleanUpActions}
\title{Perform clean-up actions}
\usage{
doCleanUpActions(x, recursive = TRUE, silent = TRUE)
}
\arguments{
\item{x}{A character vector containing paths to files or directories to be removed.}

\item{recursive}{Boolean. Should directories be deleted recursively?}

\item{silent}{Boolean. Should errors should be printed?}
}
\description{
Remove temporary files or directories. Useful for cleaning files that result from image 
segmentation procedures during optimization.
}
\details{
By default the \code{force} parameter in \code{\link[base]{unlink}} function 
(used to remove folders) is set to TRUE
}
