% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_CheckOptimFuncParameters.R
\name{doInputVerification}
\alias{doInputVerification}
\title{Checks if input parameters are valid}
\usage{
doInputVerification(input, what)
}
\arguments{
\item{input}{The input value of the parameter to check.}

\item{what}{A string defining which parameter to check. Available options are:
\itemize{
    \item "clAlgo" - classification algorithm;
    \item "classifierParams" - classification parameters;
    \item "segmMethod" - segmentation method;
    \item "evalMethod" - classifier evaluation methods;
    \item "evalMetr_sc" - evaluation metrics (single-class);
    \item "evalMetr_mc" - evaluation metric (multi-class);
    \item "dataBalancing" - data balancing methods;
    \item "trainThresh" - minimum proportion cover of a segment that decides the train class;
    \item "trainPerc" - proportion of data used for classifier training (only applies to HOCV);
    \item "segmStatsFuns" - checks if the functions used for generating segment statistics actually exist; 
}}
}
\value{
A boolean. TRUE if OK, FALSE if not
}
\description{
Validate options and ranges in user-defined inputs. This is meant to validate the parameters of 
\code{\link{gaOptimizeSegmentationParams}} general optimization function.
}
