% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_PerformanceEval.R
\name{evalPerformanceClassifier}
\alias{evalPerformanceClassifier}
\title{A function for calculating performance metrics for SOptim.Classifier objects}
\usage{
evalPerformanceClassifier(obj)
}
\arguments{
\item{obj}{An object of class \code{SOptim.Classifier} generated by \code{\link{calibrateClassifier}} 
with option \code{runFullCalibration = TRUE}.}
}
\value{
A numeric matrix with the evaluation metrics. For single-class problems an additional column 
by metric (with suffix "_Thresh") is added to the matrix with threshold values used to maximize each 
evaluation metric.
}
\description{
This function can be used to calculate all available  performance evaluation metrics both 
for single- or multi-class problems using a \code{SOptim.Classifier} objects (check out 
\code{\link{calibrateClassifier}} for more details).
}
\details{
List of available evaluation metrics:           
\strong{--- SINGLE-CLASS ---}               
\itemize{
    \item \strong{Kappa} - Cohen Kappa (maximum)
    \item \strong{PSS} - Peirce Skill score (maximum)
    \item \strong{GSS} - Gerrity Skill score (maximum)
    \item \strong{AUC} - Area Under the Receiver Operating Curve
}
\strong{--- MULTI-CLASS ---}         
\itemize{
    \item \strong{Kappa} - Cohen Kappa
    \item \strong{PSS} - Peirce Skill score
    \item \strong{GSS} - Gerrity Skill score
    \item \strong{Accuracy} - Overall accuracy
}
}
