% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_PerformanceEval.R
\name{evalPerformanceGeneric}
\alias{evalPerformanceGeneric}
\title{A generic function for calculating performance metrics}
\usage{
evalPerformanceGeneric(obs, pred, stat, nClassType)
}
\arguments{
\item{obs}{Integer vector with observed values (class labels).}

\item{pred}{Integer vector with predicted values (class labels) or numeric vector with 
predicted probabilities.}

\item{stat}{(Character) defining the statistic name ("Kappa", "PSS", "GSS", "AUC" - 
single-class only or "Accuracy" - multi-class only).}

\item{nClassType}{(Character) defining the type of classification problem, either: 
"single-class" or "multi-class".}
}
\value{
Numeric with the performance metric value
}
\description{
A single wrapper function used to calculate several performance evaluation metrics both 
for single- or multi-class problems.
}
\examples{
obs<-sample(1:5, 100, replace = TRUE)
pred<-sample(1:5, 100, replace = TRUE)
evalPerformanceGeneric(obs, pred, stat = "Kappa", nClassType = "multi-class")

}
