% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_CheckOptimFuncParameters.R
\name{evaluationMethods}
\alias{evaluationMethods}
\title{Get data balancing options}
\usage{
evaluationMethods()
}
\value{
A vector with codes/acronyms of data balancing options.
}
\description{
Outputs the codes/acronyms of data balancing options.
}
