% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_CheckOptimFuncParameters.R
\name{evaluationMetrics_multiClass}
\alias{evaluationMetrics_multiClass}
\title{Get evaluation metrics (multi-class)}
\usage{
evaluationMetrics_multiClass()
}
\value{
A vector with codes/acronyms of evaluation metrics (multi-class).
}
\description{
Outputs the codes/acronyms of evaluation metrics (multi-class).
}
