% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_CheckOptimFuncParameters.R
\name{evaluationMetrics_singleClass}
\alias{evaluationMetrics_singleClass}
\title{Get evaluation metrics (single-class)}
\usage{
evaluationMetrics_singleClass()
}
\value{
A vector with codes/acronyms of evaluation metrics (single-class).
}
\description{
Outputs the codes/acronyms of evaluation metrics (single-class).
}
