% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_AuxFunctions.R
\name{f2int}
\alias{f2int}
\title{Factor to integer conversion}
\usage{
f2int(x)
}
\arguments{
\item{x}{A factor.}
}
\value{
Returns an integer number.
}
\description{
Converts a factor variable to an integer number
}
