% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_ClassificationFunctions.R
\name{generateDefaultClassifierParams}
\alias{generateDefaultClassifierParams}
\title{Generates default parameters for classification algorithms}
\usage{
generateDefaultClassifierParams(x)
}
\arguments{
\item{x}{The dataset used for classification (this is required to calculate some classifier hyperparameters based on the number of 
columns/variables in the data).}
}
\value{
A nested list object of class \code{classificationParamList} with parameters for the available 
algorithms, namely:
         
         \itemize{
           
           \item RF - Random Forest parameters:
             \itemize{
               \item \emph{mtry} is equal to \code{floor(sqrt(ncol(x)-2))} and defines the number of variables randomly sampled 
               as candidates at each split
               \item \emph{ntree} equals 250 (by default) and is the number of trees to grow
             }
         
           \item KNN - K-nearest neighbour parameters:
             \itemize{
               \item \emph{k} is equal to 5 and is the number of neighbours considered
             }
             
           \item FDA - Flexible Discriminant Analysis with MDA-MARS parameters:
             \itemize{
               \item \emph{degree} equals 1 defining an optional integer specifying maximum interaction degree 
               \item \emph{penalty} is equal to 2 and sets an optional value specifying the cost per degree of freedom charge
               \item \emph{prune} is set to TRUE and defines an optional logical value specifying whether the model should be 
               pruned in a backward stepwise fashion
             }
             
           \item SVM - Support Vector Machine (with radial-basis kernel) parameters:
             \itemize{
               \item \emph{gamma} equals \code{1/(ncol(x)-2)} and sets the parameter needed for all kernels except linear
               \item \emph{cost} equal to 1 defines the cost of constraints violation - it is the 'C'-constant of the regularization 
               term in the Lagrange formulation
               \item \emph{probability} is equal to TRUE and defines the output type
             }
         
           \item GBM - Generalized Boosted Modeling parameters:
             \itemize{
               \item \emph{n.trees} set to 250 defining the total number of trees to fit
               \item \emph{interaction.depth} equal to 1 which defines he maximum depth of variable interactions. 1 implies an 
               additive model, 2 implies a model with up to 2-way interactions, etc  
               \item \emph{shrinkage} set to 0.01 is a shrinkage parameter applied to each tree in the expansion. Also known as the 
               learning rate or step-size reduction.
               \item \emph{bag.fraction} set to 0.5 and equals the fraction of the training set observations randomly selected to 
               propose the next tree in the expansion)
               \item \emph{distribution} set to bernoulli (if single-class) or multinomial (if multi-class) this parameter defines 
               the applicable distribution used for classification
             }
             
         }
}
\description{
This is an auxiliary function used for generating a list of default parameters used for the available 
classification algorithms: Random Forests (RF), K-nearest neighbour (KNN), Flexible Discriminant 
Analysis (FDA), Support Vector Machines (SVM) and Generalized Boosted Model (GBM).
}
\examples{

DF <- data.frame(SID=1:5, train=sample(0:1,5,replace=TRUE), Var_1=rnorm(5), Var_2=rnorm(5))
generateDefaultClassifierParams(DF)


}
\seealso{
\code{\link{replaceDefaultClassificationParams}}
}
