% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_ClassificationFunctions.R
\name{generateFormula}
\alias{generateFormula}
\title{Generate formula}
\usage{
generateFormula(x)
}
\arguments{
\item{x}{A calibration dataset with the first two columns containing the segment ID (SID) and train/test 
inputs (named as train). The remaining columns will be used as classification features.}
}
\value{
A \code{formula} object with a similar structure to: \code{train ~ feature_1 + ... + feature_n}.
}
\description{
Generates a formula object to use in classification algorithms.
}
\note{
By default removes the first 2 columns corresponding to the segment identifier (SID) and and the train columns.
}
\examples{

DF <- data.frame(SID=1:10, train = 1:10, x1 = rnorm(10), x2 = rnorm(10), x3 = rnorm(10))
generateFormula(DF)

}
