% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_PerformanceEval.R
\name{GerritySkillScore}
\alias{GerritySkillScore}
\title{Calculate the Gerrity Skill score}
\usage{
GerritySkillScore(obs, pred)
}
\arguments{
\item{obs}{Integer vector with observed values (class labels).}

\item{pred}{Integer vector with predicted values (class labels).}
}
\value{
Numeric. The GSS value calculated.
}
\description{
A  function used to calculate the Gerrity skill score (GSS). This statistic uses all entries 
in the confusion matrix, so it does not depend on the forecast distribution, and is equitable 
(i.e., random and constant forecasts score a value of 0). GSS does not reward conservative forecasting 
but rather rewards forecasts for correctly predicting the less likely categories. Smaller errors are 
penalized less than larger forecast errors. This is achieved through the use of the scoring matrix.
}
\details{
Adapted from verification package at 
URL: \url{https://github.com/cran/verification/blob/32e864afca8895391f179c2a97707e126ca31f20/R/multi.cont.R}
}
\examples{

obs<-sample(c(1:5),100,replace = TRUE)
pred<-sample(c(1:5),100,replace = TRUE)
GerritySkillScore(obs, pred)


}
