% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_AuxFunctions.R
\name{getPythonFile}
\alias{getPythonFile}
\title{Find python scripts location}
\usage{
getPythonFile(pyScriptName, pkgName, altLocs)
}
\arguments{
\item{pyScriptName}{The name of the python script.}

\item{pkgName}{The name of the package to look in.}

\item{altLocs}{Alternative locations to look for the script.}
}
\value{
If the py script is found in any location it returns the complete path.
}
\description{
Finds the path to python scripts used to perform image segmentation in ArcGIS or 
RSGISLib based on the package location or other user-defined places.
}
\note{
The \code{altLocs} parameter is used to input development paths.
}
