% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_PerformanceEval.R
\name{gssSingleClass}
\alias{gssSingleClass}
\title{Calculate the maximum Gerrity skill score}
\usage{
gssSingleClass(obs, pred)
}
\arguments{
\item{obs}{Integer vector with observed values (class labels, tipically 0 or 1).}

\item{pred}{Numeric vector with predicted probabilities.}
}
\value{
A list with two elements:
\itemize{
  \item values A data frame containing the thresholds and the calculated GSS values; 
  \item maxGSS The maximum GSS value.
}
}
\description{
A  function used to calculate the maximum value of the Gerrity skill score (GSS) 
when the predicted output is given in probability. This function applies only to 
single-class problems (e.g., presence/absence of a vegetation type, species).
}
\details{
This function will calculate the GSS value from 0 to 1 with intervals 
of 0.01. The maximum value is returned.      
This statistic uses all entries in the confusion matrix, so it does not depend 
on the forecast distribution, and is equitable (i.e., random and constant forecasts 
score a value of 0). 
GSS does not reward conservative forecasting but rather rewards forecasts for correctly 
predicting the less likely categories. Smaller errors are penalized less than 
larger forecast errors. This is achieved through the use of the scoring matrix.
}
\examples{
obs<-sample(c(0,1),100,replace = TRUE)
pred<-runif(100,0,1)
gssSingleClass(obs,pred)

}
