% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_ImportConvertData.R
\name{importToGRASS}
\alias{importToGRASS}
\title{Import files to GRASS GIS}
\usage{
importToGRASS(
  fileList,
  GRASS.path = "grass70",
  GRASS.GISDBASE,
  GRASS.LOCATION_NAME,
  GRASS.MAPSET = "PERMANENT",
  addParams = "",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{fileList}{A list of complete raster file paths to import to GRASS.}

\item{GRASS.path}{Defines the initialization path to GRASS .bat file for windows (for example: `C:/GRASS/grass70`) 
or the .sh or .py file for Linux (usually this should be set to `/usr/bin/grass`). 
The default is "grass70" which assumes that GRASS is in the path environmental variables and version 
7.0.x (which may not be appropriate in most cases and therefore this parameter should be set explicitly).}

\item{GRASS.GISDBASE}{Name of GRASS GIS database.}

\item{GRASS.LOCATION_NAME}{Name of GRASS location name (default: "demolocation").}

\item{GRASS.MAPSET}{Name of GRASS mapset (default: "PERMANENT").}

\item{addParams}{Additional parameters or flags passed to GRASS GIS \code{r.import()} (default: "" - empty string).}

\item{verbose}{Print messages? (default: TRUE)}

\item{...}{Further arguments passed to \code{\link[base]{shell}}.}
}
\description{
An auxiliary function used to import raster files into GRASS GIS using \code{r.import} command line interface. Inside R 
the \code{\link[base]{shell}} function is used to run the temporary a GRASS batch run file which will invoke a single batch 
file with \code{r.import} command.
}
\details{
By default the name of raster files imported to GRASS will have the same file basename as in \code{fileList} 
without the extension (check \code{\link[base]{basename}} and \code{\link[tools]{file_path_sans_ext}}).        
         
The working directory will be used to place some batch temporary files responsible to run GRASS GIS raster 
importer through \code{r.import()} (for more info check: \url{https://grass.osgeo.org/grass72/manuals/r.import.html}).
 
By default the function will overwrite (flag: \code{--overwrite}) any existing file with the same output name and use 
1024MB of memory (parameter: \code{memory=1024}).
}
