% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_RandSearchAlgorithm.R
\name{iterateFitnessFunc}
\alias{iterateFitnessFunc}
\title{Iterate the fitness function}
\usage{
iterateFitnessFunc(fitFunc, paramGrid, parallel = FALSE, ...)
}
\arguments{
\item{fitFunc}{A fitness/objective function object to iterate (the first input of 
this function must be a vector of parameters to be tested (named \code{x}), following 
the order of elements in a parameter list).}

\item{paramGrid}{A matrix object containing random parameters to be tested generated 
by \link{drawRandom}.}

\item{parallel}{Use parallel loop? (default: FALSE). Uses \%dorng\% as the foreach backend.}

\item{...}{Additional parameters passed to the fitness function.}
}
\value{
A vector with the fitness/objective function values. In case of error it will return \code{NA} as 
output.
}
\description{
A wrapper function used to iterate an input fitness/objective function in 
sequential or parallel form. Uses as input a matrix of parameters which represent 
points randomly drawn for a certain range of parameter values.
}
