% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_PerformanceEval.R
\name{kappaSingleClass}
\alias{kappaSingleClass}
\title{Calculate the maximum Cohen Kappa index}
\usage{
kappaSingleClass(obs, pred)
}
\arguments{
\item{obs}{Integer vector with observed values (class labels, tipically 0 or 1).}

\item{pred}{Numeric with predicted probabilities.}
}
\value{
A list with two elements:
\itemize{
  \item values A data frame containing the thresholds and the calculated Kappa values; 
  \item maxKappa The maximum Kappa value.
}
}
\description{
A  function used to calculate the maximum value of the Cohen Kappa index 
when the predicted output is given in probability. This function applies only to 
single-class problems (e.g., presence/absence of a vegetation type, species).
}
\details{
This function will calculate the Kappa value from 0 to 1 with intervals 
of 0.01. The maximum value is returned.
}
\examples{

obs<-sample(c(0,1),100,replace = TRUE)
pred<-runif(100,0,1)
kappaSingleClass(obs,pred)

}
