% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_UnsupervisedClassification.R
\name{numSampPerStrata}
\alias{numSampPerStrata}
\title{Determine and (randomly) adjust the number of samples per stratum}
\usage{
numSampPerStrata(propStrata, n, minSamp = TRUE, minSizeSet = 10)
}
\arguments{
\item{propStrata}{A \code{table} object with the proportion of each stratum 
in terms of area of relative frequency.}

\item{n}{Total number of samples.}

\item{minSamp}{Should a minimum sample size per stratum be enforced? (default: TRUE)}

\item{minSizeSet}{Minimum sample size per stratum (default: 10).}
}
\value{
A \code{table} object with the number of samples by stratum.
}
\description{
Allows to calculate the number of samples per stratum with allocation 
proportional to size
}
\examples{

st <- sample(1:5,1000,replace=TRUE)
tb <- table(st)/length(st)

numSampPerStrata(propStrata=tb, n=100, minSamp = TRUE, minSizeSet = 10)

}
\seealso{
\code{\link[base]{table}}
}
