% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_PrepFullDataset.R
\name{outputSOptimCalData}
\alias{outputSOptimCalData}
\alias{summary.SOptim.CalData}
\alias{print.SOptim.CalData}
\alias{head.SOptim.CalData}
\alias{tail.SOptim.CalData}
\title{Basic output functions for SOptim.CalData calibration data objects}
\usage{
summary.SOptim.CalData(object, ...)

print.SOptim.CalData(x, ...)

\method{head}{SOptim.CalData}(x, n = 6L, ...)

\method{tail}{SOptim.CalData}(x, n = 6L, ...)
}
\arguments{
\item{...}{Additional arguments to pass to print functions.}

\item{x, object}{An object of class \code{SOptim.CalData}.}

\item{n}{a single integer. If positive, size for the resulting object: number 
of rows for a matrix or data frame. If negative, all but the n last/first number 
of elements of x (default: 6L).}
}
\description{
Provides a set of generic output functions for summary, print, head and 
tail methods.
}
