% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_ClassificationFunctions.R
\name{predict.knn}
\alias{predict.knn}
\title{Predict k-nearest neighbour output}
\usage{
\method{predict}{knn}(
  object,
  type = "response",
  newdata = NULL,
  traindata = NULL,
  cl = NULL,
  k = NULL,
  balanceTrainData = FALSE,
  balanceMethod = "ubOver",
  ...
)
}
\arguments{
\item{object}{An object generated from \link[class]{knn} function.}

\item{type}{Type of response generated, if \code{"prob"} the probabilistic output for all classes 
will be returned. If \code{"response"} then the class labels are returned.}

\item{newdata}{A dataset used to predict the class labels.}

\item{traindata}{Train data used to calibrate the \link[class]{knn} classifier.}

\item{cl}{Factor of true classifications of training set.}

\item{k}{Number of neighbours considered.}

\item{balanceTrainData}{Defines if data balancing is to be used (only available for single-class problems; default: TRUE).}

\item{balanceMethod}{A character string used to set the data balancing method. Available methods are based on under-sampling 
\code{"ubUnder"} or over-sampling \code{"ubOver"} the target class.}

\item{...}{Additional arguments passed to \link[class]{knn}.}
}
\value{
A matrix object with knn output.
}
\description{
An auxiliary function used to predict the probabilistic outcome from the knn classifier for 
single or multi-class problems.
}
