% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_ClassificationFunctions.R
\name{print.SOptim.ClassifierSummary}
\alias{print.SOptim.ClassifierSummary}
\title{Print classification summary}
\usage{
\method{print}{SOptim.ClassifierSummary}(x)
}
\arguments{
\item{x}{\code{SOptim.ClassifierSummary} generated by 
\code{\link{summary.SOptim.Classifier}}.}
}
\value{
None
}
\description{
Prints an object of class \code{SOptim.ClassifierSummary} generated by 
\code{\link{summary.SOptim.Classifier}}.
}
