% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_PerformanceEval.R
\name{pssSingleClass}
\alias{pssSingleClass}
\title{Calculate the maximum Peirce skill score (true skill statistic or Hanssen and 
Kuipers discriminant)}
\usage{
pssSingleClass(obs, pred)
}
\arguments{
\item{obs}{Integer vector with observed values (class labels, tipically 0 or 1).}

\item{pred}{Numeric vector with predicted probabilities.}
}
\value{
A list with two elements:
\itemize{
  \item values A data frame containing the thresholds and the calculated PSS values; 
  \item maxPSS The maximum PSS value.
}
}
\description{
A  function used to calculate the maximum value of the Peirce skill score (PSS)
when the predicted output is given in probability. This function applies only to 
single-class problems (e.g., presence/absence of a vegetation type, species).
}
\details{
This function will calculate the PSS value from 0 to 1 with intervals 
of 0.01. The maximum value is returned.
}
\examples{
obs<-sample(c(0,1),100,replace = TRUE)
pred<-runif(100,0,1)
pssSingleClass(obs,pred)

}
