% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_AuxFunctions.R
\name{randString}
\alias{randString}
\title{Generate a random string}
\usage{
randString(x = 10)
}
\arguments{
\item{x}{The size of the random string to be generated (10 by default).}
}
\value{
A character string.
}
\description{
An ancillary function used to generate a random string containing integers and lettters with 
user-defined length.
}
\examples{
randString(12)

}
