% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_RasterTiling.R
\name{readDataByTile}
\alias{readDataByTile}
\title{Read raster data for a certain tile}
\usage{
readDataByTile(rst, tiles, nd = 2, tileIndex = NULL, as.df = TRUE)
}
\arguments{
\item{rst}{A \code{Raster*} object.}

\item{tiles}{A \code{SOptim.Tiles} object (generated by \code{\link{createRasterTiles}}. If not 
provided, a \code{SOptim.Tiles} will be generated on-the-fly with nd^2 tiles.}

\item{nd}{Number of times to slice the RasterLayer across row 
and column direction. The total number of tiles will be given by: 
\eqn{N_{tiles} = nd^{2}}. (default: 2)}

\item{tileIndex}{Tile index to read (default: NULL)}

\item{as.df}{Return a data.frame object? (default: TRUE)}
}
\value{
A matrix or data.frame containing pixel-level data for the whole tile. If the input is 
a \code{RasterStack} then each column represents each layer.
}
\description{
An ancilary function that uses an \code{SOptim.Tiles} object (generated by 
[createRasterTiles()]) and reads pixel-level data to a matrix or data.frame.
}
