% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_AuxFunctions.R
\name{repBSlash}
\alias{repBSlash}
\title{Replace backslash}
\usage{
repBSlash(x)
}
\arguments{
\item{x}{A string to replace.}
}
\value{
A string with replaced backslashes.
}
\description{
Replace back-slashes by forwardslashes which is probably useful for system paths
}
\examples{
repBSlash("\\\\aa\\\\bb\\\\cc")

}
