% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_ClassificationFunctions.R
\name{replaceDefaultClassificationParams}
\alias{replaceDefaultClassificationParams}
\title{Replace classification parameters}
\usage{
replaceDefaultClassificationParams(oldParamSet, newParamSet)
}
\arguments{
\item{oldParamSet}{A parameter list object of class \code{classificationParamList}.}

\item{newParamSet}{A nested list with names equal to the parameters defined in \code{\link{generateDefaultClassifierParams}}.}
}
\value{
A nested list object of class \code{classificationParamList} with parameters for the available algorithms.
}
\description{
An auxiliary function to replace default parameters of classification algorithms.
}
\examples{

# x is the classification dataset with one column with the segment ID, 
# other for train/test data inputs, and the remaining columns with classification features

x<-data.frame(SID=1:10, train=sample(c(0,1),10, replace=TRUE), feature1=rnorm(10), 
feature2=rnorm(10,5,2), feature3=rnorm(10,12,3))

oldParams<-generateDefaultClassifierParams(x)

newParams<-list(RF=list(mtry=3),
               GBM=list(n.trees=3000))

replaceDefaultClassificationParams(oldParams, newParams)

}
\seealso{
\code{\link{generateDefaultClassifierParams}} to get the list of parameters that can be changed for each 
classification algorithm.
}
