% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_SegmentationFunctions.R
\name{segmentation_GRASS_RG}
\alias{segmentation_GRASS_RG}
\title{GRASS Region Growing image segmentation}
\usage{
segmentation_GRASS_RG(
  x,
  GRASS.path = "grass",
  GRASS.inputRstName,
  GRASS.GISDBASE,
  GRASS.LOCATION_NAME = "demolocation",
  GRASS.MAPSET = "PERMANENT",
  outputSegmRst = NULL,
  Threshold = NULL,
  MinSize = NULL,
  memory = 1024,
  iterations = 20,
  verbose = TRUE
)
}
\arguments{
\item{x}{A vector with size two defining the values of segmentation parameters optimized by GA: 
\itemize{
  \item [1] threshold (float) - difference threshold between 0 and 1. Threshold: 0 merges only 
identical segments; threshold: 1 merges all. 
  \item [2] minimum number of cells in a segment (integer) - the final step will merge small 
 segments with their best neighbor; options: 1-100000 (default: 1).
}
If you are using this function outside an optimization context it is better to directly define the 
segmentation parameters in \code{Threshold} and \code{MinSize}.}

\item{GRASS.path}{Defines the initialization path to GRASS .bat file for windows (for example: `C:/GRASS/grass70`) 
or the .sh or .py file for Linux (usually this should be set to `/usr/bin/grass`). 
The default is "grass70" which assumes that GRASS is in the path environmental variables and version 
7.0.x (which may not be appropriate in most cases and therefore this parameter should be set explicitly).}

\item{GRASS.inputRstName}{Name of the raster dataset that will be segmented (typically a 
multi-layered raster dataset with segmentation features in each band).}

\item{GRASS.GISDBASE}{Name of GRASS GIS database.}

\item{GRASS.LOCATION_NAME}{Name of GRASS location name (default: "demolocation").}

\item{GRASS.MAPSET}{Name of GRASS mapset (default: "PERMANENT").}

\item{outputSegmRst}{A path to the output segmented image exported by GRASS \code{r.out.gdal}. The output file can 
be in any of GDAL formats  (default: \code{NULL}; in this case the segmenter will internally determine a temporary 
output folder and file path inside the working directory).}

\item{Threshold}{Difference threshold between 0 and 1. Threshold: 0 merges only identical segments; 
threshold: 1 merges all}

\item{MinSize}{Minimum number of cells in a segment (integer) - the final step will merge small 
segments with their best neighbor}

\item{memory}{Memory used to run i.segment (default: 1024Mb).}

\item{iterations}{Maximum number of iterations (default: 20).}

\item{verbose}{Output messages? (default: TRUE)}
}
\value{
A list object containing output file paths resulting from the segmentation run. 
These files will be cleaned after each GA iteration.
}
\description{
A function providing CLI access to GRASS Region Growing image segmentation algorithm 
and optimize parameters using genetic algorithms. To use this function, image features used in 
segmentation must be imported into a GRASS database (\code{\link{importToGRASS}}).
}
\details{
This function uses a batch file defining the batch job (GRASS_BATCH_JOB) and starts GRASS with the 
appropriate parametrization. The \code{i.segment} command is used to run the segmentation.    
The similarity calculation method uses the default value ("euclidean").     
See more info at GRASS manual pages online:
\code{i.segment} at \url{https://grass.osgeo.org/grass72/manuals/i.segment.html}   
Access to GRASS GIS manuals \url{https://grass.osgeo.org/documentation/manuals/}
}
