% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_SegmentationFunctions.R
\name{segmentation_OTB_LSMS}
\alias{segmentation_OTB_LSMS}
\title{Large Scale Mean Shift segmentation algorithm from Orfeo Toolbox (OTB)}
\usage{
segmentation_OTB_LSMS(
  x,
  inputRstPath,
  outputSegmRst = NULL,
  SpectralRange = NULL,
  SpatialRange = NULL,
  MinSize = NULL,
  lsms_maxiter = 10,
  tilesizex = 2500,
  tilesizey = 2500,
  otbBinPath = NULL,
  RAM = 3072,
  verbose = TRUE
)
}
\arguments{
\item{x}{A vector of size three containing the parameters that will be optimized by the 
genetic algorithms from package GA:    
\itemize{
    \item [1] Range radius defining the radius (expressed in radiometry units) in the 
    multi-spectral space (float);
    \item [2] Spatial radius of the neighborhood (float);
    \item [3] Minimum segment/region size. If, after the segmentation, a region is of 
    size lower than this criterion, 
    the region is merged with the "nearest" region (radiometrically) (integer).
}
If you are using this function outside an optimization context it is better to directly 
define the segmentation parameters in \code{SpectralRange}, \code{SpatialRange} and 
\code{MinSize}.}

\item{inputRstPath}{The input raster dataset used to perform the image segmentation using 
OTB LSMS algorithm (typically a multi-layered raster dataset with segmentation features 
in each band).}

\item{outputSegmRst}{A path to the output segmented image (default: \code{NULL}; in this 
case the segmenter will internally determine a temporary output folder and file path 
inside the working directory).}

\item{SpectralRange}{Range radius defining the radius (expressed in radiometry units) in 
the multi-spectral space (float).}

\item{SpatialRange}{Spatial radius of the neighborhood (float).}

\item{MinSize}{Minimum segment/region size.}

\item{lsms_maxiter}{Algorithm iterative scheme will stop if convergence hasn't been 
reached after the maximum number of iterations (integer).}

\item{tilesizex}{Size of tiles along the X-axis (integer).}

\item{tilesizey}{Size of tiles along the Y-axis (integer).}

\item{otbBinPath}{The system path to OTB binaries (e.g., where otbcli_MeanShiftSmoothing 
is located). By default this is equal to \code{NULL} which means that the system path is 
used.}

\item{RAM}{Available memory for processing the image data (in MB).}

\item{verbose}{Print output messages (default: TRUE).}
}
\value{
A list object containing output file paths resulting from the segmentation run. 
These files will be cleaned after each GA iteration or if an error occurs. Also, notice 
that some auxiliary files will be created: otb_filt_range_, otb_filt_spatial_, 
otb_segm_init_.
}
\description{
A function providing CLI access to OTB's LSMS segmentation and optimize its parameters 
using genetic algorithms. This function uses the same values for the spectral and spatial 
range parameters in the mean-shift smoothing and large-scale segmentation step. If you wish 
to use different values in each step use function \code{segmentation_OTB_LSMS2}.
}
\note{
Changed system() option to \code{show.output.on.console = TRUE} to avoid for errors 
happening on windows-10...

Check [OTB documentation](https://www.orfeo-toolbox.org/CookBook/Applications/app_LargeScaleMeanShift.html) 
for more info.
}
\references{
J. Michel, D. Youssefi and M. Grizonnet, 2015. Stable Mean-Shift Algorithm and Its 
Application to the Segmentation of Arbitrarily Large Remote Sensing Images. IEEE 
Transactions on Geoscience and Remote Sensing, 53: 2, 952-964.
}
