% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_SegmentationFunctions.R
\name{segmentation_OTB_LSMS2}
\alias{segmentation_OTB_LSMS2}
\title{Large Scale Mean Shift segmentation algorithm from Orfeo Toolbox (OTB) with two sets 
of parameters}
\usage{
segmentation_OTB_LSMS2(
  x,
  inputRstPath,
  outputSegmRst = NULL,
  MS_SpectralRange = NULL,
  MS_SpatialRange = NULL,
  LSS_SpectralRange = NULL,
  LSS_SpatialRange = NULL,
  MinSize = NULL,
  lsms_maxiter = 10,
  tilesizex = 2500,
  tilesizey = 2500,
  otbBinPath = NULL,
  RAM = 3072,
  verbose = TRUE
)
}
\arguments{
\item{x}{A vector of size five containing the parameters that will be optimized by the 
genetic algorithms from package GA:    
\itemize{
    \item [1] Range radius defining the radius (expressed in radiometry units) in the 
    multi-spectral space for the mean-shift smoothing step (float);
    \item [2] Spatial radius of the neighborhood for mean-shift smoothing (float);
    \item [3] Range radius defining the radius (expressed in radiometry units) in the 
    multi-spectral space for the large-scale segmentation step (float);
    \item [4] Spatial radius of the neighborhood for large-scale segmentation (float);
    \item [5] Minimum segment/region size. If, after the segmentation, a region is of 
    size lower than this criterion, the region is merged with the "nearest" region 
    (radiometrically) (integer).
}
If you are using this function outside an optimization context it is better to directly 
define the segmentation parameters in \code{MS_SpectralRange}, \code{MS_SpatialRange} 
\code{LSS_SpectralRange}, \code{LSS_SpatialRange} and \code{MinSize}.}

\item{inputRstPath}{The input raster dataset used to perform the image segmentation using 
OTB LSMS algorithm (typically a multi-layer raster dataset with segmentation features in 
each band).}

\item{outputSegmRst}{A path to the output segmented image (default: \code{NULL}; in this 
case the segmenter will internally determine a temporary output folder and file path 
inside the working directory).}

\item{MS_SpectralRange}{Threshold on spectral signature euclidean distance (expressed in 
radiometry unit) to consider neighborhood pixel for averaging. Higher values will be less 
edge-preserving (more similar to simple average in neighborhood), whereas lower values 
will result in less noise smoothing. Note that this parameter has no effect on processing 
time (float).}

\item{MS_SpatialRange}{Radius of the spatial neighborhood for averaging. Higher values 
will result in more smoothing and higher processing time (float).}

\item{LSS_SpectralRange}{Threshold on spectral signature euclidean distance (expressed 
in radiometry unit) to consider pixels in the same segment. A good value is half the 
range radius used in the MeanShiftSmoothing application (ranger parameter; float).}

\item{LSS_SpatialRange}{Threshold on Spatial distance to consider pixels in the same 
segment. A good value is half the spatial radius used in the MeanShiftSmoothing 
application (spatialr parameter; float).}

\item{MinSize}{Minimum segment/region size.}

\item{lsms_maxiter}{Algorithm iterative scheme will stop if convergence hasn't been 
reached after the maximum number of iterations (integer).}

\item{tilesizex}{Size of tiles along the X-axis (integer).}

\item{tilesizey}{Size of tiles along the Y-axis (integer).}

\item{otbBinPath}{The system path to OTB binaries (e.g., where otbcli_MeanShiftSmoothing 
is located). By default this is equal to \code{NULL} which means that the system path is 
used.}

\item{RAM}{Available memory for processing the image data (in MB).}

\item{verbose}{Print output messages (default: TRUE).}
}
\value{
A list object containing output file paths resulting from the segmentation run. 
These files will be cleaned after each GA iteration or if an error occurs. Also, notice 
that some auxiliary files will be created: otb_filt_range_, otb_filt_spatial_, 
otb_segm_init_.
}
\description{
A function providing CLI access to OTB's LSMS segmentation and optimize its parameters 
using genetic algorithms. This function differs from \code{segmentation_OTB_LSMS} by 
implementing two sets of parameters for the spectral and spatial range, the first for 
running mean-shift smoothing (MS_ prefix) and, the second, for the large-scale 
segmentation step (LSS_ prefix).
}
\note{
Check [OTB documentation](https://www.orfeo-toolbox.org/CookBook/Applications/app_LargeScaleMeanShift.html) 
for more info.
}
\references{
J. Michel, D. Youssefi and M. Grizonnet, 2015. Stable Mean-Shift Algorithm and Its 
Application to the Segmentation of Arbitrarily Large Remote Sensing Images. IEEE 
Transactions on Geoscience and Remote Sensing, 53: 2, 952-964.
}
