% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_SegmentationFunctions.R
\name{segmentation_RSGISLib_Shep}
\alias{segmentation_RSGISLib_Shep}
\title{Shepherd segmentation algorithm from RSGISLib}
\usage{
segmentation_RSGISLib_Shep(
  x,
  inputRstPath,
  outputSegmRst = NULL,
  NumClust = NULL,
  MinSize = NULL,
  SpectralThresh = NULL,
  pythonPath = NULL,
  verbose = TRUE
)
}
\arguments{
\item{x}{A vector of size three containing the parameters that will be optimized by the genetic algorithms 
from package GA: 
\itemize{
  \item [1] Number of clusters for running the k-means algorithm (integer)
  \item [2] Minimum segment size, if smaller it will be merged to neighbouring segments (in pixels; integer)
  \item [3] Spectral threshold used to merge objects (float), is a value providing the maximum (Euclidean 
distance) spectral separation for which to merge clumps. Set to a large value to ignore spectral 
separation and always merge.
}
If you are using this function outside an optimization context it is better to directly define the 
segmentation parameters in \code{NumClust}, \code{MinSize} and \code{SpectralThresh}.}

\item{inputRstPath}{Path to the input raster file (typically a multi-layered raster dataset with 
segmentation features in each band).}

\item{outputSegmRst}{A path to the output segmented image (default: \code{NULL}; in this case the segmenter will 
internally determine a temporary output folder and file path inside the working directory)}

\item{NumClust}{Number of clusters for running the k-means algorithm (integer)}

\item{MinSize}{Minimum segment size (in pixels; integer)}

\item{SpectralThresh}{Spectral threshold used to merge objects (float)}

\item{pythonPath}{Path to the Python interpreter (by default uses the one in the PATH env var). In windows it 
should be used for defining the path to the conda distribution of RSGISLib.}

\item{verbose}{Print output messages (default: TRUE).}
}
\value{
A list object containing output file paths resulting from the segmentation run. 
These files will be cleaned after each GA iteration.
}
\description{
A function providing CLI access to RSGISLib Shepherd segmentation and optimize its parameters 
using genetic algorithms.
}
\details{
By default the k-means algorithm running inside the segmenter will use \code{"INITCLUSTER_DIAGONAL_FULL_ATTACH"} 
for the initialization stage, 250 iterations and 0.0025 for the degree of change stopping condition.
}
\note{
Unfortunetelly these functionalities where not yet tested in Linux or Mac environments.
}
\references{
Clewley, D., Bunting, P., Shepherd, J., Gillingham, S., Flood, N., Dymond, J., Lucas, R., 
Armston, J., Moghaddam, M., 2014. A Python-Based Open Source System for Geographic 
Object-Based Image Analysis (GEOBIA) Utilizing Raster Attribute Tables. Remote 
Sensing 6, 6111.
}
\seealso{
Also check out the RSGISLib manual pages at: \url{http://www.rsgislib.org/rsgislib_segmentation.html}.
}
