% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_SegmentationFunctions.R
\name{segmentationGeneric}
\alias{segmentationGeneric}
\title{A generic dispatcher/wrapper function for running multiple segmentation methods}
\usage{
segmentationGeneric(segmentMethod, ...)
}
\arguments{
\item{segmentMethod}{Character string used to define the segmentation method. 
Available options are:     
\itemize{
  \item \code{"SAGA_SRG"} - SAGA Simple Region Growing;
  \item \code{"GRASS_RG"} - GRASS Region Growing;
  \item \code{"ArcGIS_MShift"} - ArcGIS Mean Shift algorithm;
  \item \code{"Terralib_Baatz"} - TerraLib Baatz algorithm;
  \item \code{"Terralib_MRGrow"} - TerraLib Mean Region Growing;
  \item \code{"RSGISLib_Shep"} - RSGISLib Shepherd algorithm;
  \item \code{"OTB_LSMS"} - OTB Large Scale Mean Shift algorithm;
  \item \code{"OTB_LSMS2"} - OTB Large Scale Mean Shift algorithm 
  with two separate sets of parameters, one for mean-shift smoothing 
  and another for large-scale segmentation step;
}}

\item{...}{Specific parameters to pass to each segmentation method.}
}
\description{
Runs the appropriate segmentation method with the parameters defined in \code{...}.
}
\seealso{
Check out more information about each specific segmentation method at: 
\itemize{
    \item \code{\link{segmentation_SAGA_SRG}} 
    \item \code{\link{segmentation_GRASS_RG}} 
    \item \code{\link{segmentation_ArcGIS_MShift}} 
    \item \code{\link{segmentation_Terralib_Baatz}} 
    \item \code{\link{segmentation_Terralib_MRGrow}}
    \item \code{\link{segmentation_RSGISLib_Shep}}
    \item \code{\link{segmentation_OTB_LSMS}}
    \item \code{\link{segmentation_OTB_LSMS2}}
}
}
