% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_CheckOptimFuncParameters.R
\name{segmentationMethods}
\alias{segmentationMethods}
\title{Get segmentation algorithm codes}
\usage{
segmentationMethods()
}
\value{
A vector with codes/acronyms of segmentation algorithms.
}
\description{
Outputs the codes/acronyms of segmentation algorithms.
}
