% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_SegmentationFunctions.R
\name{segmentationParamNames}
\alias{segmentationParamNames}
\title{Names of segmentation parameters that will be optimized}
\usage{
segmentationParamNames()
}
\value{
Returns the parameter names as \code{list} object.
}
\description{
Outputs a list containing the names of segmentation parameters that will be optimized by 
the package for each available segmentation algorithm: 
\itemize{
   \item SAGA_SRG - SAGA Seeded Region Growing, 
   \item GRASS_RG - GRASS Region Growing, 
   \item ArcGIS_MShift - Mean Shift algorithm from ArcGIS, 
   \item TerraLib_Baatz and TerraLib_MRGrow - Baatz algorithm and Mean Region Growing from TerraLib,  
   \item RSGISLib_Shep - Shepherd's algorithm from RSGISLib,and,
   \item Orfeo Toolbox (OTB) - Large-scale Mean-shift (single and two parameter set)
   
}
}
\seealso{
\itemize{
   \item \code{\link{segmentation_SAGA_SRG}}, 
   \item \code{\link{segmentation_GRASS_RG}}, 
   \item \code{\link{segmentation_ArcGIS_MShift}}, 
   \item \code{\link{segmentation_Terralib_Baatz}}, 
   \item \code{\link{segmentation_Terralib_MRGrow}}, 
   \item \code{\link{segmentation_RSGISLib_Shep}},
   \item \code{\link{segmentation_OTB_LSMS}},
   \item \code{\link{segmentation_OTB_LSMS2}}
}
}
