\name{SegOptim-package}

\alias{SegOptim-package}

\alias{SegOptim}

\docType{package}

\title{
\packageTitle{SegOptim}
}

\description{
\packageDescription{SegOptim}
}

\details{

The DESCRIPTION file:
\packageDESCRIPTION{SegOptim}

}

\author{
Author: \packageAuthor{SegOptim}
Maintainer: \packageMaintainer{SegOptim}
}

\references{
Scrucca L. (2013). GA: A Package for Genetic Algorithms in R. Journal of Statistical Software, 53(4), 1-37, \url{http://www.jstatsoft.org/v53/i04/}.         
             
Scrucca L. (2016). On some extensions to GA package: hybrid optimisation, parallelisation and islands evolution. Submitted to R Journal. Pre-print available at: \url{http://arxiv.org/abs/1605.01931}.
}


\keyword{Image segmentation}
\keyword{Genetic algorithms}

\seealso{
  \link{GA}
}

