% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_SimulateTestData-v1.R
\name{simRasterSegments}
\alias{simRasterSegments}
\title{Simulate raster segments}
\usage{
simRasterSegments(
  nr = 100,
  nc = 100,
  nClasses = 5,
  p = 0.3,
  ai = rep(1/nClasses, nClasses),
  neighbourhood = 4,
  directions = 8
)
}
\arguments{
\item{nr}{Number of rows for the simulated \code{RasterLayer} (default: 100).}

\item{nc}{Number of columns for the simulated \code{RasterLayer} (default: 100).}

\item{nClasses}{Number of classes to simulate. Must be equal to the length of \code{ai}}

\item{p}{[\code{numerical(1)}]\cr
Defines the proportion of elements randomly selected to form
clusters.}

\item{ai}{Vector with the cluster type distribution (percentages of occupancy).
This directly controls the number of types via the given length.}

\item{neighbourhood}{[\code{numerical(1)}]\cr
Clusters are defined using a set of neighbourhood structures,
 4 (Rook's or von Neumann neighbourhood) (default), 8 (Queen's or Moore neighbourhood).}

\item{directions}{Integer. Which cells are considered adjacent? Should be 8 (Queen's case) or 4 (Rook's case) }
}
\value{
A simulated \code{RasterLayer} with segments.
}
\description{
Uses functionalities from package NLMR to generate raster segments 
very similar to an actual image segmentation algorithm outcome.
}
