% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_SimulateTestData-v1.R
\name{simRasterSegments2}
\alias{simRasterSegments2}
\title{Simulate raster segments (faster version)}
\usage{
simRasterSegments2(
  nr = 100,
  nc = 100,
  nClasses = 10,
  winSize = 3,
  directions = 8
)
}
\arguments{
\item{nr}{Number of rows for the simulated \code{SpatRaster} (default: 100).}

\item{nc}{Number of columns for the simulated \code{SpatRaster} (default: 100).}

\item{nClasses}{Number of classes to simulate (default: 4). Larger numbers will 
tend to generate more segments.}

\item{winSize}{Window size for performing the focal operation using the modal function 
(higher values will generate less segments; default: 3)}

\item{directions}{integer indicating which cells are considered adjacent. Should be 8 (Queen's case) or 4 (Rook's case)}
}
\value{
A simulated \code{SpatRaster} with segments.
}
\description{
Uses simulated data for n classes, performs modal smoothing and then 
checks for raster clumps similar to segments
}
