% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_SimulateTestData-v1.R
\name{simRasterTrain}
\alias{simRasterTrain}
\title{Simulate raster train data}
\usage{
simRasterTrain(
  nr = 100,
  nc = 100,
  classes = c(0, 1, NA),
  probs = c(0.1, 0.2, 0.7)
)
}
\arguments{
\item{nr}{Number of rows for the simulated \code{RasterLayer} (default: 100).}

\item{nc}{Number of columns for the simulated \code{RasterLayer} (default: 100).}

\item{classes}{A vector defining integer train classes (default for 
single-class: \code{c(0,1,NA)} but can be used for multi-class; 
\code{NA} should always be included to include background nodata).}

\item{probs}{A vector defining sampling probabilities for each class to use in 
[base::sample()] (default: \code{c(0.1,0.2,0.7)}). Values sum up to one if 
not are adjusted by sample function internally.}
}
\value{
A \code{RasterLayer} object with random train data.
}
\description{
Uses random data to create a simulated \code{RasterLayer} useful 
for testing and debugging code.
}
