% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_UnsupervisedClassification.R
\name{StRS}
\alias{StRS}
\title{Performs Stratified Random Sampling}
\usage{
StRS(x, strata, nsps)
}
\arguments{
\item{x}{Vector of integer indices}

\item{strata}{A vector containing the stratum number/code for each element of \code{x} 
(unique elements in strata must equal the names property of \code{nsps}).}

\item{nsps}{An object of class \code{table} indicating the number of samples per stratum (check 
\link{numSampPerStrata}).}
}
\value{
A vector of indices with the selected elements of \code{x}.
}
\description{
An auxiliar function to perform stratified random sampling of a set of integers.
}
\examples{

st <- sample(1:5,1000,replace=TRUE)
tb <- table(st)/length(st)
ind <- 1:1000

nsps <- numSampPerStrata(propStrata=tb, n=100, minSamp = TRUE, minSizeSet = 10)

StRS(x = ind, strata = st, nsps = nsps)

}
\seealso{
\code{\link[base]{sample}}
}
