% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOptim_ClassificationFunctions.R
\name{summary.SOptim.Classifier}
\alias{summary.SOptim.Classifier}
\title{Summary for SOptim.Classifier objects}
\usage{
\method{summary}{SOptim.Classifier}(object, ...)
}
\arguments{
\item{object}{An object of class \code{SOptim.Classifier} created by 
\code{\link{calibrateClassifier}} with option \code{runFullCalibration=TRUE}.}

\item{...}{Additional arguments affecting the summary produced.}
}
\value{
A summary of the object of class \code{SOptim.ClassifierSummary} 
which can be used in print
}
\description{
Produces a summary of classification performance results for a 
\code{SOptim.Classifier} object.
}
\seealso{
\link{calibrateClassifier}
}
